orthographic projection

+++ include/const.h
@@ -1,7 +1,7 @@
 #ifndef CONST_H
 #define CONST_H
 
-const int SCREEN_WIDTH = 800;
-const int SCREEN_HEIGHT = 800;
+const float SCREEN_WIDTH = 800.0f;
+const float SCREEN_HEIGHT = 800.0f;
 
 #endif
